Fix(Archives download): Will not download the same archive twice

+++ screens/install/install.py
@@ -354,15 +354,17 @@ class   Install:
         for conf in lst:
             dl_ok = 0
             i = 0
-            # Test multiples urls in case one fail
-            # TODO: Handle exception from retrieving
-            while dl_ok == 0:
-                # Update the progress bar
-                self.dlg.gauge_update(int((to_dl * 100) / dl_len),
-                    "Getting "+ conf["archive"]+ "... ("+ str(to_dl) +"/"+ str(dl_len) +")", True)
-                urlretrieve(conf["urls"][i], self.arch_dir + conf["archive"])
-                i += 1
-                dl_ok = 1
+            if os.path.isfile(self.arch_dir + conf["archive"] == False):
+                # Test multiples urls in case one fail
+                # TODO: Handle exception from retrieving
+                while dl_ok == 0:
+                    # Update the progress bar
+                    self.dlg.gauge_update(int((to_dl * 100) / dl_len),
+                        "Getting "+ conf["archive"]+ "... ("+ str(to_dl) +"/"+
+                        str(dl_len) +")", True)
+                    urlretrieve(conf["urls"][i], self.arch_dir + conf["archive"])
+                    i += 1
+                    dl_ok = 1
             to_dl += 1
         # Stop the progress bar
         self.dlg.gauge_stop()
